+++ test/mvc_application_test.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:mvc_application/app.dart';
+//import 'package:flutter_test/flutter_test.dart';
+//
+//import 'package:mvc_application/view.dart';
 
 void main() {}